(chore) move all sign-up components to a sign-up folder
