+++ src/client/components/app.routes.ts
@@ -3,7 +3,7 @@ import { MenuBarComponent } from './menuBar.component'
 import { AdvancedSearch } from './advancedSearch.component';
 import { StudentDashboard } from './studentDashboard.component';
 import { Login } from './login.component';
-import { SignUpComponent } from './signUp.component';
+import { SignUpComponent } from './sign-up/signUp.component';
 import { TeacherDashboard } from './teacherDashboard.component';
 import { App } from '../containers/app.component';
