+++ src/client/components/sign-up/signUp.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { NgRedux } from 'ng2-redux';
-import { IAppState, rootReducer, enhancers } from '../store/index';
+import { IAppState, rootReducer, enhancers } from '../../store/index';
 import { User } from './user';
 import { UserFormComponent } from './user-form.component';
